+++ Modules/Intent.Modules.Eventing.Contracts.DomainMapping/Templates/MessageExtensions/MessageExtensionsTemplatePartial.cs
@@ -1,14 +1,17 @@
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using Intent.Engine;
 using Intent.Eventing.Contracts.DomainMapping.Api;
 using Intent.Metadata.Models;
 using Intent.Modelers.Domain.Api;
+using Intent.Modelers.Domain.Events.Api;
 using Intent.Modelers.Eventing.Api;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
+using Intent.Modules.Constants;
 using Intent.Modules.Eventing.Contracts.Templates.IntegrationEventDto;
 using Intent.Modules.Eventing.Contracts.Templates.IntegrationEventMessage;
 using Intent.RoslynWeaver.Attributes;
@@ -20,7 +23,7 @@ using Intent.Templates;
 namespace Intent.Modules.Eventing.Contracts.DomainMapping.Templates.MessageExtensions
 {
     [IntentManaged(Mode.Fully, Body = Mode.Merge)]
-    partial class MessageExtensionsTemplate : CSharpTemplateBase<MessageModel>, ICSharpFileBuilderTemplate
+    public partial class MessageExtensionsTemplate : CSharpTemplateBase<MessageModel>, ICSharpFileBuilderTemplate
     {
         public const string TemplateId = "Intent.Eventing.Contracts.DomainMapping.MessageExtensions";
 
@@ -29,8 +32,9 @@ namespace Intent.Modules.Eventing.Contracts.DomainMapping.Templates.MessageExten
         {
             AddTypeSource(IntegrationEventMessageTemplate.TemplateId);
             AddTypeSource(IntegrationEventDtoTemplate.TemplateId);
-            AddTypeSource("Domain.Enum");
-            AddTypeSource("Domain.Entity");
+            AddTypeSource(TemplateFulfillingRoles.Domain.Enum);
+            AddTypeSource(TemplateFulfillingRoles.Domain.Entity.Primary);
+            AddTypeSource(TemplateFulfillingRoles.Domain.Events);
 
             CSharpFile = new CSharpFile($"{Model.InternalElement.Package.Name.ToPascalCase()}", this.GetFolderPath())
                 .AddClass($"{Model.Name.RemoveSuffix("Event")}EventExtensions", @class =>
@@ -43,13 +47,26 @@ namespace Intent.Modules.Eventing.Contracts.DomainMapping.Templates.MessageExten
                         method.AddParameter(GetTypeName(model.GetMapFromDomainMapping()), "projectFrom", param => param.WithThisModifier());
 
                         var domainEntity = ((IElement)model.GetMapFromDomainMapping().Element).AsClassModel();
+                        if (domainEntity != null)
+                        {
+                            var codeLines = new CSharpStatementAggregator();
+                            codeLines.Add($"return new {GetTypeName(model.InternalElement)}");
+                            codeLines.Add(new CSharpStatementBlock()
+                                .AddStatements(MappingExtensionHelper.GetPropertyAssignments("projectFrom", domainEntity.Attributes.Select(x => x.InternalElement), model.Properties.Select(x => x.InternalElement), this))
+                                .WithSemicolon());
+                            method.AddStatements(codeLines.ToList());
+                        }
 
-                        var codeLines = new CSharpStatementAggregator();
-                        codeLines.Add($"return new {GetTypeName(model.InternalElement)}");
-                        codeLines.Add(new CSharpStatementBlock()
-                            .AddStatements(MappingExtensionHelper.GetPropertyAssignments("projectFrom", domainEntity, model.Properties.Select(x => x.InternalElement), this))
-                            .WithSemicolon());
-                        method.AddStatements(codeLines.ToList());
+                        var domainEvent = ((IElement)model.GetMapFromDomainMapping().Element).AsDomainEventModel();
+                        if (domainEvent != null)
+                        {
+                            var codeLines = new CSharpStatementAggregator();
+                            codeLines.Add($"return new {GetTypeName(model.InternalElement)}");
+                            codeLines.Add(new CSharpStatementBlock()
+                                .AddStatements(MappingExtensionHelper.GetPropertyAssignments("projectFrom", domainEvent.Properties.Select(x => x.InternalElement), model.Properties.Select(x => x.InternalElement), this))
+                                .WithSemicolon());
+                            method.AddStatements(codeLines.ToList());
+                        }
                     });
                 });
         }