+++ Modules/Intent.Modules.Eventing.Contracts.DomainMapping/Templates/TemplateExtensions.cs
@@ -12,21 +12,22 @@ namespace Intent.Modules.Eventing.Contracts.DomainMapping.Templates
 {
     public static class TemplateExtensions
     {
-        public static string GetDtoExtensionsName<T>(this IntentTemplateBase<T> template) where T : Intent.Modelers.Eventing.Api.EventingDTOModel
+        public static string GetDtoExtensionsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Eventing.Api.EventingDTOModel
         {
             return template.GetTypeName(DtoExtensionsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetDtoExtensionsName(this IntentTemplateBase template, Intent.Modelers.Eventing.Api.EventingDTOModel model)
+        public static string GetDtoExtensionsName(this IIntentTemplate template, Intent.Modelers.Eventing.Api.EventingDTOModel model)
         {
             return template.GetTypeName(DtoExtensionsTemplate.TemplateId, model);
         }
-        public static string GetMessageExtensionsName<T>(this IntentTemplateBase<T> template) where T : Intent.Modelers.Eventing.Api.MessageModel
+
+        public static string GetMessageExtensionsName<T>(this IIntentTemplate<T> template) where T : Intent.Modelers.Eventing.Api.MessageModel
         {
             return template.GetTypeName(MessageExtensionsTemplate.TemplateId, template.Model);
         }
 
-        public static string GetMessageExtensionsName(this IntentTemplateBase template, Intent.Modelers.Eventing.Api.MessageModel model)
+        public static string GetMessageExtensionsName(this IIntentTemplate template, Intent.Modelers.Eventing.Api.MessageModel model)
         {
             return template.GetTypeName(MessageExtensionsTemplate.TemplateId, model);
         }