+++ Modules/Intent.Modules.MediatR.DomainEvents/Templates/DomainEventNotification/DomainEventNotificationTemplateRegistration.cs
@@ -8,7 +8,7 @@ using Intent.Modules.Common.Registrations;
 using Intent.RoslynWeaver.Attributes;
 using Intent.Templates;
 
-[assembly: DefaultIntentManaged(Mode.Merge)]
+[assembly: DefaultIntentManaged(Mode.Fully)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.TemplateRegistration.SingleFileNoModel", Version = "1.0")]
 
 namespace Intent.Modules.MediatR.DomainEvents.Templates.DomainEventNotification
@@ -20,7 +20,7 @@ namespace Intent.Modules.MediatR.DomainEvents.Templates.DomainEventNotification
 
         public override ITemplate CreateTemplateInstance(IOutputTarget outputTarget)
         {
-            return new DomainEventNotificationTemplate(outputTarget, null);
+            return new DomainEventNotificationTemplate(outputTarget);
         }
     }
 }