+++ Modules/Intent.Modules.Eventing.Contracts.DomainMapping/release-notes.md
@@ -1,3 +1,7 @@
+### Version 4.1.0
+
+- Now supports mapping Messages from Domain Events and will automatically implement publishing in the default Domain Event Handler.
+
 ### Version 4.0.3
 
 - Updated dependencies and supported client versions to prevent warnings when used with Intent Architect 4.x.